Muchos cambios

+++ app/src/main/java/com/example/smarthometec/ui/profile/ProfileFragment.java
@@ -1,6 +1,8 @@
 package com.example.smarthometec.ui.profile;
 
+import android.app.Activity;
 import android.content.Intent;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -19,10 +21,19 @@ import com.example.smarthometec.MainActivity;
 import com.example.smarthometec.Menu;
 import com.example.smarthometec.R;
 import com.example.smarthometec.Register;
+import com.example.smarthometec.UserDeviceInputs;
 import com.example.smarthometec.ui.database.Aposento;
 import com.example.smarthometec.ui.database.DatabaseHandler;
 import com.example.smarthometec.ui.database.Dispositivo;
 import com.example.smarthometec.ui.database.User;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
 
 
 public class ProfileFragment extends Fragment {
@@ -57,14 +68,61 @@ public class ProfileFragment extends Fragment {
         deleteacc.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                User user = db.getUser(m.email);
-                db.deleteAposento(new Aposento("", user.getEmail()));
-                db.deleteDispositivo(new Dispositivo(user.getEmail()));
-                db.deleteUser(user);
-                Toast.makeText(getContext(), "Su cuenta se ha eliminado exitosamente.", Toast.LENGTH_SHORT).show();
-                Intent i = new Intent(getActivity(),MainActivity.class);
-                startActivity(i);
-                getActivity().finish();
+                AsyncTask.execute(new Runnable() {//CORREGIR
+                    @Override
+                    public void run() {
+                        try {
+                            URL url = new URL("http://192.168.0.14/api/login/BorrarUsuario");
+
+                            HttpURLConnection con = (HttpURLConnection) url.openConnection();
+                            con.setRequestMethod("POST");
+
+                            con.setRequestProperty("Content-Type", "application/json; utf-8");
+                            con.setRequestProperty("Accept", "application/json");
+
+                            con.setDoOutput(true);
+
+                            String jsonInputString = "{\"Nombre\":" +   "null"  + "," +
+                                    "\"Apellido\":" +  "null" + "," +
+                                    "\"Correo\":" + "\"" + m.email + "\"" + "," +
+                                    "\"Contraseña\":" + "null"  + "," +
+                                    "\"Direccion\":" + "null"  + "," +
+                                    "\"Continente\":" + "null"  + "," +
+                                    "\"Pais\":" + "null"  + "}";
+
+                            try (OutputStream os = con.getOutputStream()) {
+                                byte[] input = jsonInputString.getBytes("utf-8");
+                                os.write(input, 0, input.length);
+                            }
+
+                            int code = con.getResponseCode();
+                            System.out.println(code);
+
+                            try (BufferedReader br = new BufferedReader(new InputStreamReader(con.getInputStream(), "utf-8"))) {
+                                StringBuilder response = new StringBuilder();
+                                String responseLine = null;
+                                while ((responseLine = br.readLine()) != null) {
+                                    response.append(responseLine.trim());
+                                }
+                                System.out.println(response.toString());
+                                if(response.toString().equals("\"El usuario ha sido eliminado exitosamente\"")){
+                                    db.deleteAposento(new Aposento("", user.getEmail()));
+                                    db.deleteDispositivo(new Dispositivo(user.getEmail()));
+                                    db.deleteUser(user);
+                                    Toast.makeText(getContext(), "Su cuenta se ha eliminado exitosamente.", Toast.LENGTH_SHORT).show();
+                                    Intent i = new Intent(getActivity(),MainActivity.class);
+                                    startActivity(i);
+                                    getActivity().finish();
+                                }
+                            }
+                        } catch (MalformedURLException e) {
+                            e.printStackTrace();
+                        } catch (IOException e) {
+                            e.printStackTrace();
+                        }
+                    }
+                });
+
             }
         });
         return root;
